+++ include/spt/area_light.h
@@ -11,8 +11,8 @@ class AreaLight : public Light
 public:
     AreaLight(const Ref<Primitive> primitive);
 
-    virtual Color Sample(Vec3* wi, f64* pdf, f64* visibility, const Intersection& ref) const override;
-    virtual f64 EvaluatePDF(const Ray& ray) const override;
+    virtual Color Sample(Vec3* wi, Float* pdf, Float* visibility, const Intersection& ref) const override;
+    virtual Float EvaluatePDF(const Ray& ray) const override;
 
     const Primitive* GetPrimitive() const;
 
@@ -20,7 +20,7 @@ private:
     Ref<Primitive> primitive;
 };
 
-inline f64 AreaLight::EvaluatePDF(const Ray& ray) const
+inline Float AreaLight::EvaluatePDF(const Ray& ray) const
 {
     return primitive->EvaluatePDF(ray);
 }
