+++ include/spt/camera.h
@@ -13,12 +13,12 @@ public:
     Camera(const Point3& look_from,
            const Point3& look_at,
            const Vec3& up,
-           f64 vfov, // vertical field-of-view. in degrees.
-           f64 aspect_ratio,
-           f64 aperture,
-           f64 focus_dist);
+           Float vfov, // vertical field-of-view. in degrees.
+           Float aspect_ratio,
+           Float aperture,
+           Float focus_dist);
 
-    Ray GetRay(f64 s, f64 t) const;
+    Ray GetRay(Float s, Float t) const;
 
     Point3 origin;
     Vec3 dir;
@@ -27,7 +27,7 @@ private:
     Point3 lower_left;
     Vec3 horizontal, vertical;
 
-    f64 lens_radius;
+    Float lens_radius;
 
     // Local coordinate frame
     Vec3 u, v, w;
