+++ include/spt/cosine_pdf.h
@@ -12,7 +12,7 @@ public:
     CosinePDF(const Vec3& n);
 
     virtual Vec3 Sample() const override;
-    virtual f64 Evaluate(const Vec3& wi) const override;
+    virtual Float Evaluate(const Vec3& wi) const override;
 
 public:
     ONB uvw;
@@ -28,10 +28,10 @@ inline Vec3 CosinePDF::Sample() const
     return uvw.GetLocal(CosineSampleHemisphere());
 }
 
-inline f64 CosinePDF::Evaluate(const Vec3& wi) const
+inline Float CosinePDF::Evaluate(const Vec3& wi) const
 {
-    f64 cosine = Dot(wi, uvw.w);
-    return cosine <= 0.0 ? 0.0 : cosine * inv_pi;
+    Float cosine = Dot(wi, uvw.w);
+    return cosine <= 0 ? 0 : cosine * inv_pi;
 }
 
 } // namespace spt
