+++ include/spt/directional_light.h
@@ -8,17 +8,17 @@ namespace spt
 struct DirectionalLight : public Light
 {
 public:
-    DirectionalLight(const Vec3& dir, const Color& intensity, f64 radius);
+    DirectionalLight(const Vec3& dir, const Color& intensity, Float radius);
 
-    virtual Color Sample(Vec3* wi, f64* pdf, f64* visibility, const Intersection& ref) const override;
-    virtual f64 EvaluatePDF(const Ray& ray) const override;
+    virtual Color Sample(Vec3* wi, Float* pdf, Float* visibility, const Intersection& ref) const override;
+    virtual Float EvaluatePDF(const Ray& ray) const override;
 
     Vec3 dir;
     Color intensity; // radiance
-    f64 radius;      // visible radius
+    Float radius;      // visible radius
 };
 
-inline f64 DirectionalLight::EvaluatePDF(const Ray& ray) const
+inline Float DirectionalLight::EvaluatePDF(const Ray& ray) const
 {
     assert(false);
     return 0.0;
